seperate out styling in masterstyles
