+++ src/Components/AbilityScores/AbilityScores.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { formatModifier, modifiers } from "../CharacterSheet/utils";
 import { useRollFunctions } from "../utils/diceRoller";
 import { useTheme } from "../../contexts/ThemeContext";
-import { getAbilityScoresStyles } from "../../styles/masterStyles";
+import { getAbilityScoresStyles } from "./styles";
 
 const AbilityScores = ({ character }) => {
   const { rollAbility, rollSavingThrow } = useRollFunctions();
