+++ src/Components/CharacterNotes/styles.js
@@ -1,0 +1,371 @@
+export const createCharacterNotesStyles = (theme) => ({
+  container: {
+    margin: "0 auto",
+    padding: "20px",
+  },
+
+  header: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: "20px",
+    padding: "20px",
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
+  },
+  headerContent: {
+    display: "flex",
+    flexDirection: "column",
+  },
+  title: {
+    margin: "0 0 8px 0",
+    fontSize: "28px",
+    color: theme.text,
+    display: "flex",
+    alignItems: "center",
+    gap: "12px",
+  },
+  subtitle: {
+    margin: "0",
+    color: theme.text,
+    fontSize: "16px",
+  },
+  headerActions: {
+    display: "flex",
+    gap: "12px",
+    alignItems: "center",
+  },
+  newEntryButton: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    padding: "12px 20px",
+    backgroundColor: theme.primary,
+    color: "white",
+    border: "none",
+    borderRadius: "8px",
+    fontSize: "16px",
+    fontWeight: "600",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+  },
+
+  noCharacterContainer: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    justifyContent: "center",
+    padding: "60px 20px",
+    backgroundColor: "#fef2f2",
+    border: "2px solid #fecaca",
+    borderRadius: "12px",
+    margin: "20px",
+    maxWidth: "600px",
+    marginLeft: "auto",
+    marginRight: "auto",
+  },
+  noCharacterIcon: {
+    marginBottom: "16px",
+  },
+  noCharacterTitle: {
+    color: theme.text,
+    margin: "0 0 8px 0",
+    fontSize: "24px",
+  },
+  noCharacterText: {
+    color: theme.text,
+    margin: "0",
+    textAlign: "center",
+  },
+  newEntryForm: {
+    backgroundColor: theme.surface,
+    padding: "20px",
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    marginBottom: "20px",
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
+  },
+  newEntryTitle: {
+    margin: "0 0 12px 0",
+    fontSize: "18px",
+    color: theme.text,
+  },
+  newEntryInput: {
+    width: "100%",
+    padding: "12px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "8px",
+    fontSize: "16px",
+    outline: "none",
+    backgroundColor: theme.background,
+    color: theme.text,
+    marginBottom: "16px",
+  },
+  templateButtons: {
+    display: "flex",
+    gap: "8px",
+    flexWrap: "wrap",
+    marginBottom: "16px",
+  },
+  templateButton: {
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+    padding: "8px 12px",
+    backgroundColor: theme.background,
+    border: `1px solid ${theme.border}`,
+    borderRadius: "6px",
+    fontSize: "14px",
+    cursor: "pointer",
+    color: theme.text,
+    transition: "all 0.2s ease",
+  },
+  formActions: {
+    display: "flex",
+    gap: "8px",
+  },
+  cancelButton: {
+    padding: "8px 16px",
+    backgroundColor: theme.background,
+    border: `1px solid ${theme.border}`,
+    borderRadius: "6px",
+    cursor: "pointer",
+    color: theme.text,
+  },
+  editForm: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
+    overflow: "hidden",
+  },
+  editHeader: {
+    padding: "16px 20px",
+    borderBottom: `1px solid ${theme.border}`,
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  editTitleInput: {
+    flex: 1,
+    padding: "8px 12px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "6px",
+    fontSize: "18px",
+    fontWeight: "600",
+    marginRight: "12px",
+    backgroundColor: theme.background,
+    color: theme.text,
+  },
+  editActions: {
+    display: "flex",
+    gap: "8px",
+  },
+  saveButton: {
+    padding: "8px 16px",
+    backgroundColor: theme.success + "20",
+    border: `1px solid ${theme.success}`,
+    borderRadius: "6px",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    color: theme.success,
+    fontSize: "14px",
+    fontWeight: "600",
+    gap: "6px",
+  },
+  editCancelButton: {
+    padding: "8px 16px",
+    backgroundColor: theme.background,
+    border: `1px solid ${theme.border}`,
+    borderRadius: "6px",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    fontSize: "14px",
+    fontWeight: "600",
+    gap: "6px",
+    color: theme.text,
+  },
+  templateSection: {
+    padding: "16px 20px",
+    borderBottom: `1px solid ${theme.border}`,
+    backgroundColor: theme.background,
+  },
+  templateSectionTitle: {
+    margin: "0 0 12px 0",
+    fontSize: "14px",
+    color: theme.text,
+    fontWeight: "600",
+  },
+  editorContainer: {
+    padding: "20px",
+  },
+  editorLoading: {
+    height: "500px",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    color: theme.textSecondary,
+  },
+  entriesGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fill, minmax(400px, 1fr))",
+    gap: "20px",
+  },
+  entryCard: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
+    overflow: "hidden",
+    transition: "all 0.2s ease",
+  },
+  entryHeader: {
+    padding: "16px 20px",
+    borderBottom: `1px solid ${theme.border}`,
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "flex-start",
+  },
+  entryTitleSection: {
+    flex: 1,
+  },
+  entryTitle: {
+    margin: "0 0 4px 0",
+    fontSize: "18px",
+    color: theme.text,
+    fontWeight: "600",
+  },
+  entryDate: {
+    margin: "0",
+    fontSize: "12px",
+    color: theme.text,
+  },
+  entryActions: {
+    display: "flex",
+    gap: "8px",
+    marginLeft: "12px",
+  },
+  editEntryButton: {
+    padding: "6px",
+    backgroundColor: theme.secondary,
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    color: theme.text,
+    transition: "all 0.2s ease",
+  },
+  deleteEntryButton: {
+    padding: "6px",
+    backgroundColor: theme.warning,
+    border: "1px solid #fecaca",
+    borderRadius: "4px",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    color: theme.text,
+    transition: "all 0.2s ease",
+  },
+  duplicateEntryButton: {
+    padding: "6px",
+    backgroundColor: theme.primary,
+    border: `1px solid #457880`,
+    borderRadius: "4px",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    color: theme.text,
+    transition: "all 0.2s ease",
+  },
+  entryContent: {
+    padding: "20px",
+  },
+
+  loadingContainer: {
+    padding: "60px 20px",
+    textAlign: "center",
+    color: theme.textSecondary,
+    fontSize: "16px",
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+  },
+  emptyContainer: {
+    padding: "60px 20px",
+    textAlign: "center",
+    color: theme.textSecondary,
+    fontSize: "16px",
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+  },
+
+  importExportSection: {
+    backgroundColor: theme.surface,
+    padding: "16px 20px",
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    marginTop: "20px",
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
+  },
+  importExportActions: {
+    display: "flex",
+    gap: "8px",
+    alignItems: "center",
+  },
+  exportButton: {
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+    padding: "8px 12px",
+    backgroundColor: theme.success + "20",
+    border: `1px solid ${theme.success}`,
+    borderRadius: "6px",
+    fontSize: "14px",
+    cursor: "pointer",
+    color: theme.success,
+  },
+  importLabel: {
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+    padding: "8px 12px",
+    backgroundColor: theme.primary + "20",
+    border: `1px solid ${theme.primary}`,
+    borderRadius: "6px",
+    fontSize: "14px",
+    cursor: "pointer",
+    color: theme.primary,
+  },
+  hiddenFileInput: {
+    display: "none",
+  },
+
+  templateButtonHover: {
+    backgroundColor: theme.primary + "10",
+    borderColor: theme.primary,
+  },
+  editEntryButtonHover: {
+    backgroundColor: theme.primary + "20",
+    borderColor: theme.primary,
+  },
+  deleteEntryButtonHover: {
+    backgroundColor: "#fee2e2",
+    borderColor: "#dc2626",
+  },
+  newEntryButtonHover: {
+    backgroundColor: theme.secondary,
+    transform: "translateY(-1px)",
+    boxShadow: `0 4px 12px ${theme.primary}40`,
+  },
+});
+
+// Helper function for backward compatibility
+export const getCharacterNotesStyles = (theme) => createCharacterNotesStyles(theme);
