+++ src/Components/CharacterSheet/Skills/Skills.js
@@ -10,7 +10,6 @@ import {
   getPassiveSkillBreakdown,
 } from "../utils";
 import { MagicalTheoryModal } from "./MagicalTheoryModal";
-import { createThemedStyles } from "../../../styles/masterStyles";
 import { useRollFunctions } from "../../utils/diceRoller";
 import { useTheme } from "../../../contexts/ThemeContext";
 import { allSkills, skillMap, skillDescriptions } from "../../../SharedData";
