+++ src/Components/Downtime/styles.js
@@ -1,253 +1,855 @@
-export const downtimeStyles = (theme) => ({
-  container: {
-    maxWidth: "1200px",
-    margin: "0 auto",
-    padding: "1.5rem",
-    backgroundColor: theme.background,
-  },
-  section: {
-    marginBottom: "2rem",
-    padding: "1.5rem",
+export const createDowntimeStyles = (theme) => ({
+  activitiesContainer: {
     backgroundColor: theme.surface,
     borderRadius: "12px",
-    border: `1px solid ${theme.border}`,
-  },
-  sectionTitle: {
-    fontSize: "1.5rem",
-    fontWeight: "600",
-    color: theme.text,
-    marginBottom: "1rem",
-    borderBottom: `2px solid ${theme.primary}`,
-    paddingBottom: "0.5rem",
+    padding: "24px",
+    marginBottom: "24px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.1)",
   },
   activityCard: {
-    padding: "1.5rem",
     backgroundColor: theme.background,
-    borderRadius: "8px",
-    border: `1px solid ${theme.border}`,
-    marginBottom: "1rem",
+    borderRadius: "12px",
+    padding: "20px",
+    marginBottom: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.05)",
   },
-  inputGroup: {
-    marginBottom: "1rem",
+  activityHeader: {
+    marginBottom: "16px",
+    paddingBottom: "12px",
+    borderBottom: `2px solid ${theme.border}`,
   },
-  label: {
-    display: "block",
-    fontSize: "0.875rem",
+  activityTitle: {
+    fontSize: "18px",
     fontWeight: "600",
     color: theme.text,
-    marginBottom: "0.5rem",
+    margin: "0",
   },
-  input: {
-    width: "100%",
-    padding: "0.75rem",
-    borderRadius: "6px",
-    border: `1px solid ${theme.border}`,
+  activityContent: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "20px",
+  },
+  activitySection: {
     backgroundColor: theme.surface,
-    color: theme.text,
-    fontSize: "0.875rem",
-  },
-  select: {
-    width: "100%",
-    padding: "0.75rem",
-    borderRadius: "6px",
-    border: `1px solid ${theme.border}`,
-    backgroundColor: theme.surface,
-    color: theme.text,
-    fontSize: "0.875rem",
-  },
-  textarea: {
-    width: "100%",
-    padding: "0.75rem",
-    borderRadius: "6px",
-    border: `1px solid ${theme.border}`,
-    backgroundColor: theme.surface,
-    color: theme.text,
-    fontSize: "0.875rem",
-    resize: "none",
-    height: "80px",
-  },
-  activityDescription: {
-    padding: "1rem",
-    backgroundColor: theme.primary + "10",
-    border: `1px solid ${theme.primary + "30"}`,
-    borderRadius: "6px",
-    fontSize: "0.875rem",
-    color: theme.text,
-    lineHeight: "1.5",
-    marginTop: "0.5rem",
-    marginBottom: "0.5rem",
-  },
-  diceAssignment: {
-    display: "grid",
-    gridTemplateColumns: "repeat(auto-fit, minmax(200px, 1fr))",
-    gap: "1rem",
-    marginTop: "1rem",
-  },
-  assignedDice: {
-    padding: "1rem",
-    backgroundColor: theme.success + "15",
     borderRadius: "8px",
-    border: `2px solid ${theme.success}`,
-    textAlign: "center",
-    color: theme.success,
-    maxWidth: "450px",
+    padding: "16px",
+    border: `1px solid ${theme.border}`,
+  },
+  rollSection: {
+    backgroundColor: theme.surface,
+    borderRadius: "8px",
+    padding: "16px",
+    border: `1px solid ${theme.border}`,
+  },
+  rollSectionTitle: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.primary,
+    marginBottom: "12px",
+    margin: "0 0 12px 0",
+  },
+  selectedRollArea: {
+    display: "flex",
+    gap: "16px",
+    marginBottom: "12px",
+  },
+  notesGroup: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  assignedDiceContainer: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+    width: "70px",
+    marginBottom: "12px",
+  },
+  assignedDiceTile: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    gap: "8px",
+    padding: "12px",
+    backgroundColor: theme.primary + "20",
+    border: `2px solid ${theme.primary}`,
+    borderRadius: "8px",
+    transition: "all 0.2s ease",
+  },
+  diceDisplay: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    gap: "4px",
+    width: "24px",
+    height: "32px",
   },
   diceValue: {
-    fontSize: "1.5rem",
+    fontSize: "24px",
     fontWeight: "bold",
-    color: theme.success,
-    marginBottom: "0.5rem",
+    color: theme.primary,
   },
-  total: {
-    fontSize: "1rem",
-    fontWeight: "600",
-    color: theme.text,
-    marginBottom: "0.5rem",
+  diceLabel: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    fontWeight: "500",
   },
-  removeButton: {
-    padding: "0.25rem 0.75rem",
-    backgroundColor: theme.error + "15",
-    color: theme.error,
-    border: `1px solid ${theme.error}`,
-    borderRadius: "4px",
-    fontSize: "0.75rem",
-    cursor: "pointer",
-    marginTop: "0.5rem",
-  },
-  button: {
-    padding: "0.75rem 1.5rem",
-    borderRadius: "8px",
+  removeDiceButton: {
+    padding: "4px 8px",
+    backgroundColor: "#ef4444",
+    color: "white",
     border: "none",
-    fontWeight: "600",
+    borderRadius: "4px",
+    fontSize: "12px",
+    fontWeight: "500",
     cursor: "pointer",
     transition: "all 0.2s ease",
   },
-  secondaryButton: {
-    backgroundColor: theme.warning,
+  successSection: {
+    backgroundColor: theme.surface,
+    borderRadius: "8px",
+    padding: "16px",
+    border: `1px solid ${theme.border}`,
+  },
+  successHeader: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: "12px",
+  },
+  adminOnlyBadge: {
+    padding: "2px 8px",
+    backgroundColor: "#ef4444",
     color: "white",
+    borderRadius: "12px",
+    fontSize: "10px",
+    fontWeight: "600",
+    textTransform: "uppercase",
+  },
+  successGrid: {
+    display: "flex",
+    gap: "8px",
+    marginBottom: "12px",
   },
   successButton: {
-    backgroundColor: theme.success,
-    color: "white",
-  },
-  actionButtons: {
+    width: "32px",
+    height: "32px",
+    borderRadius: "50%",
+    border: "2px solid",
+    backgroundColor: "transparent",
+    color: theme.text,
+    fontSize: "12px",
+    fontWeight: "600",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
     display: "flex",
-    gap: "1rem",
+    alignItems: "center",
     justifyContent: "center",
-    paddingTop: "1rem",
   },
-  makeSpellContainer: {
-    backgroundColor: theme.surface,
-    border: `1px solid ${theme.border}`,
+  successText: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    fontWeight: "500",
+    margin: "0",
+  },
+  adminCompletedNotice: {
+    backgroundColor: `#8b5cf6${"20"}`,
+    color: "#8b5cf6",
+    padding: "16px",
     borderRadius: "8px",
-    padding: "1rem",
-    marginTop: "0.5rem",
+    border: `2px solid #8b5cf6`,
+    marginBottom: "20px",
+    textAlign: "center",
   },
-  progressContainer: {
-    marginTop: "1rem",
-    padding: "1rem",
+  completedNotice: {
+    backgroundColor: `#8b5cf6${"20"}`,
+    color: "#8b5cf6",
+    padding: "12px",
+    borderRadius: "6px",
+    textAlign: "center",
+    fontSize: "14px",
+    fontWeight: "600",
+  },
+  container: {
+    margin: "0 auto",
+    padding: "20px",
+    fontFamily: "Arial, sans-serif",
+    fontSize: "14px",
     backgroundColor: theme.background,
+    color: theme.text,
+    minHeight: "100vh",
+    maxWidth: "1200px",
+  },
+  loadingOverlay: {
+    position: "fixed",
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: "rgba(0, 0, 0, 0.5)",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    zIndex: 1000,
+  },
+  loadingSpinner: {
+    backgroundColor: theme.surface,
+    padding: "20px 40px",
+    borderRadius: "8px",
+    fontSize: "16px",
+    fontWeight: "bold",
+    color: theme.text,
+  },
+  emptyState: {
+    textAlign: "center",
+    padding: "60px 20px",
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+  },
+  formContainer: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "24px",
+    marginBottom: "24px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.1)",
+  },
+  characterInfo: {
+    marginBottom: "24px",
+    textAlign: "center",
+  },
+  title: {
+    fontSize: "28px",
+    fontWeight: "bold",
+    color: theme.text,
+    marginBottom: "8px",
+  },
+  characterDetails: {
+    display: "flex",
+    justifyContent: "center",
+    gap: "24px",
+    marginTop: "16px",
+    fontSize: "16px",
+    color: theme.textSecondary,
+  },
+  selectionSection: {
+    backgroundColor: theme.background,
+    padding: "20px",
+    borderRadius: "8px",
+    marginBottom: "24px",
+    border: `1px solid ${theme.border}`,
+  },
+  selectionRow: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr",
+    gap: "20px",
+  },
+  selectionGroup: {
+    display: "flex",
+    flexDirection: "column",
+    padding: "4px",
+    gap: "8px",
+  },
+  magicSchoolSection: {
+    backgroundColor: theme.background,
+    padding: "20px",
+    borderRadius: "8px",
+    marginBottom: "24px",
+    border: `1px solid ${theme.border}`,
+  },
+  sectionHeader: {
+    marginBottom: "16px",
+  },
+  subtitle: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    margin: "4px 0 0 0",
+  },
+  magicSchoolGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(200px, 1fr))",
+    gap: "12px",
+  },
+  magicSchoolChoice: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    padding: "8px",
+    cursor: "pointer",
+  },
+  checkboxText: {
+    fontSize: "14px",
+    color: theme.text,
+  },
+  instructionsSection: {
+    backgroundColor: theme.background,
+    padding: "20px",
+    borderRadius: "8px",
+    marginBottom: "24px",
+    border: `1px solid ${theme.border}`,
+  },
+  instructions: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+  },
+  instructionItem: {
+    padding: "12px",
+    backgroundColor: theme.surface,
+    borderRadius: "6px",
+    borderLeft: `4px solid ${theme.primary}`,
+  },
+  guidelinesSection: {
+    backgroundColor: theme.background,
+    padding: "20px",
+    borderRadius: "8px",
+    marginBottom: "24px",
+    border: `1px solid ${theme.border}`,
+  },
+  guidelines: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "20px",
+  },
+  guidelineCategory: {
+    backgroundColor: theme.surface,
+    padding: "16px",
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+  },
+  guidelineList: {
+    marginLeft: "20px",
+    marginTop: "8px",
+  },
+  diceSection: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "24px",
+    marginBottom: "24px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.1)",
+  },
+  diceHeader: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: "20px",
+  },
+  dicePool: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "16px",
+  },
+  diceGrid: {
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  diceItem: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    padding: "15px",
+    borderRadius: "8px",
+    textAlign: "center",
+    height: "60px",
+    width: "60px",
+  },
+  diceUsage: {
+    fontSize: "10px",
+    color: theme.textSecondary,
+  },
+  diceStats: {
+    display: "flex",
+    justifyContent: "space-around",
+    padding: "12px",
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    fontSize: "14px",
+    fontWeight: "bold",
+  },
+  emptyDice: {
+    textAlign: "center",
+    padding: "40px 20px",
+    color: theme.textSecondary,
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    border: `2px dashed ${theme.border}`,
+  },
+  activityRow: {
+    display: "grid",
+    gridTemplateColumns: "120px 1fr",
+    gap: "12px",
+    alignItems: "center",
+    marginBottom: "12px",
+  },
+  rollHeader: {
+    fontSize: "16px",
+    fontWeight: "bold",
+    marginBottom: "12px",
+    color: theme.primary,
+  },
+  rollRow: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr",
+    gap: "16px",
+    marginBottom: "12px",
+  },
+  diceAssignment: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  assignedDice: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    padding: "8px 12px",
+    backgroundColor: theme.background,
+    borderRadius: "4px",
+    border: `1px solid ${theme.border}`,
+  },
+  modifierSelection: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  secondRoll: {
+    marginTop: "16px",
+    paddingTop: "16px",
+    borderTop: `1px solid ${theme.border}`,
+  },
+  secondRollHeader: {
+    fontSize: "14px",
+    fontWeight: "bold",
+    marginBottom: "8px",
+    color: theme.textSecondary,
+  },
+
+  npcSection: {
+    backgroundColor: theme.surface,
+    padding: "16px",
+    borderRadius: "6px",
+    marginTop: "16px",
+    border: `1px solid ${theme.border}`,
+  },
+  npcHeader: {
+    fontSize: "16px",
+    fontWeight: "bold",
+    marginBottom: "12px",
+    color: theme.secondary || theme.primary,
+  },
+  npcRow: {
+    display: "grid",
+    gridTemplateColumns: "120px 1fr",
+    gap: "12px",
+    alignItems: "center",
+    marginBottom: "12px",
+  },
+  successCheckboxes: {
+    display: "flex",
+    gap: "8px",
+    marginTop: "8px",
+  },
+  checkboxLabel: {
+    display: "flex",
+    alignItems: "center",
+    gap: "4px",
+    cursor: "pointer",
+  },
+  successDots: {
+    display: "flex",
+    gap: "4px",
+    marginTop: "4px",
+  },
+  successDot: {
+    width: "12px",
+    height: "12px",
+    borderRadius: "50%",
+    border: `1px solid ${theme.border}`,
+  },
+
+  notesSection: {
+    marginTop: "16px",
+  },
+  resultsContainer: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "24px",
+    marginBottom: "24px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.1)",
+  },
+  resultsHeader: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: "20px",
+  },
+  summaryStats: {
+    display: "flex",
+    gap: "20px",
+  },
+  stat: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    gap: "4px",
+  },
+  statLabel: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+  },
+  statValue: {
+    fontSize: "16px",
+    fontWeight: "bold",
+    color: theme.text,
+  },
+  noResults: {
+    textAlign: "center",
+    padding: "40px 20px",
+    color: theme.textSecondary,
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    border: `2px dashed ${theme.border}`,
+  },
+  resultsGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(300px, 1fr))",
+    gap: "16px",
+  },
+
+  resultCard: {
+    backgroundColor: theme.background,
+    padding: "16px",
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+  },
+  resultHeader: {
+    marginBottom: "12px",
+  },
+  activityName: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    margin: "4px 0 0 0",
+  },
+  resultDetails: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+  },
+  rollBreakdown: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "4px",
+    fontSize: "14px",
+  },
+  diceRoll: {
+    fontWeight: "bold",
+    color: theme.text,
+  },
+  modifier: {
+    color: theme.textSecondary,
+  },
+  totalResult: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "8px 12px",
+    backgroundColor: theme.surface,
+    borderRadius: "6px",
+    fontWeight: "bold",
+  },
+  totalNumber: {
+    fontSize: "18px",
+    color: theme.text,
+  },
+  dcInfo: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+  },
+  successLevel: {
+    fontSize: "14px",
+    fontWeight: "bold",
+  },
+  notes: {
+    marginTop: "12px",
+    padding: "8px",
+    backgroundColor: theme.surface,
+    borderRadius: "4px",
+    fontSize: "12px",
+    color: theme.textSecondary,
+  },
+
+  progressSummary: {
+    marginTop: "24px",
+    padding: "20px",
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+  },
+  progressGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(250px, 1fr))",
+    gap: "16px",
+  },
+  progressItem: {
+    padding: "12px",
+    backgroundColor: theme.surface,
     borderRadius: "6px",
     border: `1px solid ${theme.border}`,
   },
   progressHeader: {
+    fontSize: "14px",
+    fontWeight: "bold",
+    color: theme.primary,
+    marginBottom: "4px",
+  },
+  progressActivity: {
+    fontSize: "12px",
+    color: theme.text,
+    marginBottom: "4px",
+  },
+  progressNpc: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    marginBottom: "8px",
+  },
+  progressSuccesses: {
+    fontSize: "12px",
+    fontWeight: "bold",
+    color: theme.text,
+  },
+
+  actionsContainer: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "24px",
+    marginBottom: "24px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 2px 4px rgba(0,0,0,0.1)",
+  },
+
+  viewModeSelector: {
+    display: "flex",
+    gap: "8px",
+    marginBottom: "20px",
+  },
+  viewModeButton: {
+    padding: "8px 16px",
+    border: "none",
+    borderRadius: "6px",
+    cursor: "pointer",
+    fontSize: "14px",
+    fontWeight: "bold",
+    transition: "all 0.2s ease",
+  },
+
+  actionButtons: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "16px",
+  },
+  buttonGroup: {
+    display: "flex",
+    gap: "12px",
+    flexWrap: "wrap",
+  },
+  button: {
+    padding: "10px 20px",
+    border: "none",
+    borderRadius: "6px",
+    cursor: "pointer",
+    fontSize: "14px",
+    fontWeight: "bold",
+    color: "white",
+    transition: "all 0.2s ease",
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+  },
+  removeButton: {
+    padding: "4px 8px",
+    backgroundColor: "#ef4444",
+    color: "white",
+    border: "none",
+    borderRadius: "4px",
+    fontSize: "12px",
+    cursor: "pointer",
+  },
+
+  submittedStatus: {
+    display: "flex",
+    alignItems: "center",
+  },
+  submittedText: {
+    color: "#10b981",
+    fontWeight: "bold",
+  },
+  statusInfo: {
+    marginTop: "16px",
+    padding: "12px",
+    backgroundColor: theme.background,
+    borderRadius: "6px",
+  },
+  draftNotice: {
+    padding: "8px 12px",
+    backgroundColor: "#fef3c7",
+    color: "#92400e",
+    borderRadius: "4px",
+    marginBottom: "8px",
+  },
+  lastEdited: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+  },
+
+  characterSummary: {
+    marginTop: "16px",
+    padding: "16px",
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+  },
+  summaryGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(150px, 1fr))",
+    gap: "8px",
+    marginTop: "8px",
+    fontSize: "14px",
+  },
+
+  label: {
+    fontSize: "14px",
+    fontWeight: "bold",
+    color: theme.text,
+    marginBottom: "4px",
+    display: "block",
+  },
+  select: {
+    width: "100%",
+    padding: "8px 12px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "14px",
+  },
+  input: {
+    width: "100%",
+    padding: "8px 12px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "14px",
+  },
+  textarea: {
+    width: "100%",
+    padding: "8px 12px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "14px",
+    minHeight: "60px",
+    resize: "vertical",
+  },
+  checkbox: {
+    width: "16px",
+    height: "16px",
+    accentColor: theme.primary,
+  },
+
+  sheet: {
+    backgroundColor: theme.surface,
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    overflow: "hidden",
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  headerSection: {
+    display: "grid",
+    gridTemplateColumns: "2fr 2fr",
+    borderBottom: `2px solid ${theme.border}`,
+  },
+  nameSection: {
+    padding: "12px",
+    borderRight: `2px solid ${theme.border}`,
+  },
+  gradesSection: {
+    padding: "12px",
+    textAlign: "center",
+    fontWeight: "bold",
+    fontSize: "14px",
+    backgroundColor: theme.primary + "20",
+    borderBottom: `2px solid ${theme.border}`,
+    color: theme.text,
+  },
+  yearSemesterSection: {
+    display: "grid",
+    gridTemplateColumns: "2fr 1fr",
+    borderBottom: `2px solid ${theme.border}`,
+  },
+  typeSection: {
+    padding: "12px",
+  },
+  semesterSection: {
+    padding: "12px",
+  },
+  subjectsGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(3, 1fr)",
+    borderBottom: `2px solid ${theme.border}`,
+  },
+  subjectColumn: {
+    borderRight: `2px solid ${theme.border}`,
+    "&:last-child": {
+      borderRight: "none",
+    },
+  },
+  subjectItem: {
     display: "flex",
     justifyContent: "space-between",
     alignItems: "center",
-    marginBottom: "1rem",
+    padding: "8px 12px",
+    borderBottom: `1px solid ${theme.border}`,
   },
-  progressTitle: {
-    margin: 0,
-    fontSize: "1.1rem",
-    fontWeight: "600",
-    color: theme.text,
+  activitiesSection: {
+    borderBottom: `2px solid ${theme.border}`,
   },
-  progressBadge: {
-    backgroundColor: theme.primary,
-    color: "white",
-    padding: "0.25rem 0.75rem",
-    borderRadius: "12px",
-    fontSize: "0.875rem",
-    fontWeight: "600",
-  },
-  checksList: {
-    display: "flex",
-    flexDirection: "column",
-    gap: "0.75rem",
-    marginBottom: "1rem",
-  },
-  checkItem: {
-    display: "flex",
-    alignItems: "flex-start",
-    gap: "0.75rem",
-    padding: "0.75rem",
-    borderRadius: "6px",
-    border: `1px solid ${theme.border}`,
-    backgroundColor: theme.surface,
-  },
-  checkItemCompleted: {
-    backgroundColor: theme.primary + "10",
-    borderColor: theme.primary,
-  },
-  checkItemCurrent: {
-    backgroundColor: theme.secondary + "10",
-    borderColor: theme.secondary,
-  },
-  checkIcon: {
-    width: "24px",
-    height: "24px",
-    borderRadius: "50%",
-    backgroundColor: theme.primary,
-    color: "white",
-    display: "flex",
+  relationshipRow: {
+    display: "grid",
+    gridTemplateColumns: "50px 1fr auto auto",
     alignItems: "center",
-    justifyContent: "center",
-    fontSize: "14px",
-    fontWeight: "bold",
-    flexShrink: 0,
+    padding: "12px",
+    borderBottom: `1px solid ${theme.border}`,
+    gap: "8px",
+    backgroundColor: theme.background,
   },
-  checkContent: {
-    flex: 1,
-  },
-  checkName: {
-    fontSize: "0.95rem",
-    fontWeight: "600",
+  relationshipInput: {
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    padding: "6px 8px",
+    backgroundColor: theme.surface,
     color: theme.text,
-    marginBottom: "0.25rem",
   },
-  checkDescription: {
-    fontSize: "0.85rem",
-    color: theme.textSecondary,
-    lineHeight: 1.4,
-  },
-  currentCheckInfo: {
-    padding: "1rem",
-    backgroundColor: theme.secondary + "10",
-    border: `1px solid ${theme.secondary}`,
-    borderRadius: "6px",
-  },
-  currentCheckTitle: {
-    fontSize: "1rem",
-    fontWeight: "600",
+  gradeInput: {
+    width: "60px",
+    padding: "4px 6px",
+    border: `1px solid ${theme.border}`,
+    borderRadius: "4px",
+    backgroundColor: theme.surface,
     color: theme.text,
-    marginBottom: "0.5rem",
-  },
-  dcInfo: {
-    fontSize: "0.95rem",
-    color: theme.text,
-    marginBottom: "0.5rem",
-  },
-  checkNote: {
-    fontSize: "0.85rem",
-    color: theme.textSecondary,
-    fontStyle: "italic",
-  },
-  completionMessage: {
-    padding: "1rem",
-    backgroundColor: theme.primary + "10",
-    border: `1px solid ${theme.primary}`,
-    borderRadius: "6px",
     textAlign: "center",
-    color: theme.text,
   },
 });
+
+// Helper function for backward compatibility
+export const getDowntimeStyles = (theme) => createDowntimeStyles(theme);
+
+// Keeping the old export for backward compatibility
+export const downtimeStyles = (theme) => createDowntimeStyles(theme);
