+++ src/Components/Potions/styles.js
@@ -1,0 +1,496 @@
+export const createPotionsStyles = (theme) => ({
+  container: {
+    maxWidth: "1400px",
+    margin: "0 auto",
+    padding: "20px",
+    backgroundColor: theme.background,
+    minHeight: "100vh",
+    fontFamily:
+      '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif',
+  },
+  title: {
+    fontSize: "32px",
+    fontWeight: "bold",
+    textAlign: "center",
+    marginBottom: "24px",
+    color: theme.text,
+    textShadow: "2px 2px 4px rgba(0,0,0,0.1)",
+  },
+  themeControls: {
+    display: "flex",
+    gap: "16px",
+    justifyContent: "center",
+    marginBottom: "24px",
+    flexWrap: "wrap",
+  },
+  themeControlGroup: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+  },
+  themeLabel: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  themeSelect: {
+    padding: "8px 12px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "6px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "14px",
+  },
+  controlsContainer: {
+    display: "flex",
+    justifyContent: "center",
+    marginBottom: "24px",
+  },
+  card: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "20px",
+    marginBottom: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  thresholdCard: {
+    display: "flex",
+    fontSize: "14px",
+    justifyContent: "space-between",
+    color: theme.text,
+    backgroundColor: theme.surface,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+
+    borderTop: `2px solid ${theme.border}`,
+    marginLeft: "4px",
+  },
+  thresholdCardTitle: {
+    paddingBottom: "6px",
+    fontWeight: "semi-bold",
+    color: theme.highlight,
+    margin: 0,
+  },
+  cardTitle: {
+    fontSize: "20px",
+    fontWeight: "bold",
+    color: theme.text,
+    marginBottom: "16px",
+    margin: 0,
+  },
+  proficiencyGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(300px, 1fr))",
+    gap: "20px",
+  },
+  proficiencySection: {
+    backgroundColor: theme.background,
+    padding: "16px",
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+    color: theme.text,
+  },
+  proficiencyTitle: {
+    fontSize: "16px",
+    fontWeight: "600",
+    color: theme.text,
+    marginBottom: "12px",
+  },
+  inputGroup: {
+    marginBottom: "12px",
+  },
+  inputLabel: {
+    display: "block",
+    fontSize: "14px",
+    fontWeight: "500",
+    color: theme.text,
+    marginBottom: "4px",
+  },
+  select: {
+    width: "100%",
+    padding: "8px 12px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "6px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "14px",
+  },
+  checkboxGroup: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  checkboxLabel: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    fontSize: "14px",
+    color: theme.text,
+    cursor: "pointer",
+  },
+  maxQualityInfo: {
+    marginTop: "12px",
+    padding: "8px 12px",
+    backgroundColor: theme.primary + "20",
+    borderRadius: "6px",
+    fontSize: "14px",
+    color: theme.text,
+    fontWeight: "500",
+  },
+  searchContainer: {
+    display: "flex",
+    gap: "12px",
+    alignItems: "center",
+    flexWrap: "wrap",
+  },
+  searchInputContainer: {
+    position: "relative",
+    flex: 1,
+    minWidth: "200px",
+  },
+  searchIcon: {
+    position: "absolute",
+    left: "12px",
+    top: "50%",
+    transform: "translateY(-50%)",
+    color: theme.textSecondary,
+  },
+  searchInput: {
+    width: "100%",
+    padding: "10px 12px 10px 40px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    backgroundColor: theme.surface,
+    color: theme.text,
+    fontSize: "16px",
+  },
+  mainGrid: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr",
+    gap: "20px",
+  },
+  potionList: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+    maxHeight: "600px",
+    overflowY: "auto",
+    padding: "4px",
+  },
+  potionItem: {
+    padding: "16px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+    backgroundColor: theme.background,
+  },
+  potionItemSelected: {
+    borderColor: theme.primary,
+    backgroundColor: theme.primary + "10",
+  },
+  potionItemHover: {
+    borderColor: theme.primary + "80",
+    backgroundColor: theme.surface,
+  },
+  potionHeader: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "flex-start",
+    marginBottom: "8px",
+  },
+  potionName: {
+    fontSize: "16px",
+    fontWeight: "600",
+    color: theme.text,
+    margin: 0,
+  },
+  potionMeta: {
+    display: "flex",
+    gap: "8px",
+    alignItems: "center",
+  },
+  potionYear: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    fontWeight: "500",
+  },
+  rarityBadge: {
+    fontSize: "10px",
+    fontWeight: "600",
+    padding: "2px 6px",
+    borderRadius: "10px",
+    textTransform: "uppercase",
+    color: "white",
+  },
+  rarityColors: {
+    common: "#6B7280",
+    uncommon: "#10B981",
+    rare: "#3B82F6",
+    "very rare": "#8B5CF6",
+    legendary: "#F59E0B",
+  },
+  potionDescription: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    lineHeight: "1.4",
+    margin: 0,
+  },
+  brewingInterface: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "16px",
+  },
+  selectedPotionCard: {
+    backgroundColor: theme.primary + "10",
+    border: `2px solid ${theme.primary}`,
+    borderRadius: "12px",
+    padding: "20px",
+  },
+  selectedPotionTitle: {
+    fontSize: "20px",
+    fontWeight: "bold",
+    color: theme.text,
+    marginBottom: "8px",
+    margin: 0,
+  },
+  selectedPotionDescription: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    lineHeight: "1.5",
+    margin: 0,
+  },
+  brewButton: {
+    backgroundColor: theme.primary,
+    color: "white",
+    border: "none",
+    borderRadius: "8px",
+    padding: "16px 24px",
+    fontSize: "18px",
+    fontWeight: "600",
+    cursor: "pointer",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: "12px",
+    transition: "all 0.2s ease",
+    width: "100%",
+  },
+  brewButtonDisabled: {
+    backgroundColor: theme.textSecondary,
+    cursor: "not-allowed",
+  },
+  resultCard: {
+    backgroundColor: theme.surface,
+    border: `2px solid ${theme.border}`,
+    borderRadius: "12px",
+    padding: "20px",
+  },
+  resultTitle: {
+    fontSize: "18px",
+    fontWeight: "600",
+    color: theme.text,
+    marginBottom: "16px",
+    margin: 0,
+  },
+  resultGrid: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+  },
+  resultRow: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  resultLabel: {
+    fontSize: "14px",
+    fontWeight: "500",
+    color: theme.textSecondary,
+  },
+  resultValue: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  diceResult: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+  },
+  diceValue: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  qualityResult: {
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+    fontSize: "14px",
+    fontWeight: "600",
+    textTransform: "capitalize",
+  },
+  qualityRuined: {
+    color: "#EF4444",
+  },
+  qualityFlawed: {
+    color: "#F59E0B",
+  },
+  qualityNormal: {
+    color: "#10B981",
+  },
+  qualityExceptional: {
+    color: "#3B82F6",
+  },
+  qualitySuperior: {
+    color: "#8B5CF6",
+  },
+  resultTimestamp: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    fontStyle: "italic",
+    textAlign: "center",
+    marginTop: "8px",
+  },
+  referenceGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(300px, 1fr))",
+    gap: "20px",
+    marginTop: "24px",
+  },
+  referenceTable: {
+    width: "100%",
+    borderCollapse: "collapse",
+  },
+  referenceTableHeader: {
+    display: "grid",
+    gridTemplateColumns: "1fr repeat(4, 0.8fr)",
+    gap: "8px",
+    padding: "12px",
+    backgroundColor: theme.background,
+    fontWeight: "600",
+    fontSize: "14px",
+    color: theme.text,
+    borderBottom: `2px solid ${theme.border}`,
+  },
+  referenceTableRow: {
+    display: "grid",
+    gridTemplateColumns: "1fr repeat(4, 0.8fr)",
+    gap: "8px",
+    padding: "8px 12px",
+    borderBottom: `1px solid ${theme.border}`,
+    fontSize: "14px",
+    color: theme.text,
+  },
+  referenceTableRowHeader: {
+    fontWeight: "600",
+    textTransform: "capitalize",
+  },
+  ingredientEffectsList: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  ingredientEffectRow: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "8px 12px",
+    backgroundColor: theme.background,
+    borderRadius: "6px",
+    fontSize: "14px",
+  },
+  ingredientEffectLabel: {
+    fontWeight: "600",
+    textTransform: "capitalize",
+    color: theme.text,
+  },
+  ingredientEffectValue: {
+    fontWeight: "500",
+    color: theme.text,
+  },
+  ingredientEffectHarder: {
+    color: "#EF4444",
+  },
+  ingredientEffectEasier: {
+    color: "#10B981",
+  },
+  button: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    padding: "10px 16px",
+    border: "none",
+    borderRadius: "6px",
+    fontSize: "14px",
+    fontWeight: "600",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+    color: "white",
+  },
+  primaryButton: {
+    backgroundColor: theme.primary,
+  },
+  ingredientTransformation: {
+    marginTop: "12px",
+    padding: "8px 12px",
+    backgroundColor: theme.surface,
+    borderRadius: "6px",
+    border: `1px solid ${theme.border}`,
+  },
+
+  transformationRow: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    marginBottom: "8px",
+  },
+
+  rawQuality: {
+    fontSize: "14px",
+    color: theme.text,
+    fontWeight: "500",
+  },
+
+  preparedQuality: {
+    fontSize: "14px",
+    color: theme.text,
+    fontWeight: "600",
+  },
+
+  dcModifier: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    fontStyle: "italic",
+  },
+
+  rollPreview: {
+    marginTop: "16px",
+  },
+
+  rollPreviewContent: {
+    padding: "8px",
+    margin: "4px",
+    backgroundColor: theme.surface,
+    borderRadius: "6px",
+  },
+
+  totalValue: {
+    fontWeight: "700",
+    color: theme.primary,
+  },
+
+  warningText: {
+    marginTop: "8px",
+    padding: "8px",
+    backgroundColor: theme.warning || "#FEF3C7",
+    color: theme.warningText || "#92400E",
+    fontSize: "14px",
+    borderRadius: "4px",
+    textAlign: "center",
+  },
+});
+
+// Helper function for backward compatibility
+export const getPotionsStyles = (theme) => createPotionsStyles(theme);
