+++ src/Components/Recipes/styles.js
@@ -1,0 +1,482 @@
+export const createRecipesStyles = (theme) => ({
+  container: {
+    maxWidth: "1400px",
+    margin: "0 auto",
+    padding: "20px",
+    backgroundColor: theme.background,
+    minHeight: "100vh",
+    fontFamily:
+      '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif',
+  },
+  header: {
+    textAlign: "center",
+    marginBottom: "32px",
+    padding: "24px",
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  headerContent: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: "16px",
+  },
+  headerIcon: {
+    color: theme.primary,
+  },
+  title: {
+    fontSize: "32px",
+    fontWeight: "bold",
+    color: theme.text,
+    margin: 0,
+  },
+  subtitle: {
+    fontSize: "16px",
+    color: theme.textSecondary,
+    margin: "8px 0 0 0",
+  },
+  content: {
+    display: "grid",
+    gridTemplateColumns: "1fr 400px",
+    gap: "24px",
+    alignItems: "start",
+  },
+  recipeLibrary: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  searchAndFilter: {
+    display: "flex",
+    gap: "16px",
+    marginBottom: "16px",
+    alignItems: "center",
+  },
+  searchContainer: {
+    position: "relative",
+    flex: 1,
+  },
+  searchIcon: {
+    position: "absolute",
+    left: "12px",
+    top: "50%",
+    transform: "translateY(-50%)",
+    color: theme.textSecondary,
+  },
+  searchInput: {
+    width: "100%",
+    padding: "12px 12px 12px 44px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    backgroundColor: theme.background,
+    color: theme.text,
+    fontSize: "14px",
+    outline: "none",
+    transition: "border-color 0.2s ease",
+  },
+  categoryFilter: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+  },
+  filterIcon: {
+    color: theme.textSecondary,
+  },
+  categorySelect: {
+    padding: "12px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    backgroundColor: theme.background,
+    color: theme.text,
+    fontSize: "14px",
+    outline: "none",
+    cursor: "pointer",
+  },
+  recipeCount: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    marginBottom: "16px",
+    textAlign: "center",
+  },
+  recipeGrid: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+    maxHeight: "600px",
+    overflowY: "auto",
+    paddingRight: "8px",
+  },
+  recipeItem: {
+    padding: "16px",
+    backgroundColor: theme.background,
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+  },
+  recipeItemSelected: {
+    backgroundColor: theme.primary + "20",
+    borderColor: theme.primary,
+    transform: "translateY(-2px)",
+    boxShadow: "0 4px 8px rgba(0, 0, 0, 0.15)",
+  },
+  recipeHeader: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "flex-start",
+    marginBottom: "8px",
+  },
+  recipeName: {
+    fontSize: "16px",
+    fontWeight: "600",
+    color: theme.text,
+    margin: 0,
+    flex: 1,
+  },
+  recipeMeta: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "4px",
+    alignItems: "flex-end",
+  },
+  recipeTime: {
+    display: "flex",
+    alignItems: "center",
+    gap: "4px",
+    fontSize: "12px",
+    color: theme.textSecondary,
+    backgroundColor: theme.surface,
+    padding: "2px 6px",
+    borderRadius: "4px",
+  },
+  recipeDuration: {
+    display: "flex",
+    alignItems: "center",
+    gap: "4px",
+    fontSize: "12px",
+    color: theme.textSecondary,
+    backgroundColor: theme.surface,
+    padding: "2px 6px",
+    borderRadius: "4px",
+  },
+  recipeDescription: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    lineHeight: "1.4",
+    margin: 0,
+  },
+  cookingInterface: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "20px",
+  },
+  selectedRecipeCard: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  selectedRecipeTitle: {
+    display: "flex",
+    alignItems: "center",
+    gap: "12px",
+    fontSize: "20px",
+    fontWeight: "600",
+    color: theme.text,
+    marginBottom: "12px",
+    margin: 0,
+  },
+  selectedRecipeDescription: {
+    fontSize: "14px",
+    color: theme.textSecondary,
+    lineHeight: "1.5",
+    marginBottom: "16px",
+  },
+  recipeDetails: {
+    backgroundColor: theme.background,
+    padding: "12px",
+    borderRadius: "8px",
+    marginBottom: "16px",
+  },
+  detailRow: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: "8px",
+  },
+  detailLabel: {
+    fontSize: "14px",
+    fontWeight: "500",
+    color: theme.textSecondary,
+  },
+  detailValue: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  rollPreview: {
+    marginBottom: "20px",
+  },
+  rollPreviewContent: {
+    marginBottom: "12px",
+  },
+  ingredientSection: {
+    marginBottom: "16px",
+  },
+  sectionTitle: {
+    fontSize: "16px",
+    fontWeight: "600",
+    color: theme.text,
+    marginBottom: "12px",
+  },
+  ingredientControls: {
+    display: "grid",
+    gridTemplateColumns: "repeat(2, 1fr)",
+    gap: "8px",
+  },
+  qualityButton: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    gap: "4px",
+    padding: "12px 8px",
+    border: `2px solid ${theme.border}`,
+    borderRadius: "8px",
+    backgroundColor: theme.background,
+    color: theme.text,
+    fontSize: "14px",
+    fontWeight: "500",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+  },
+  qualityButtonActive: {
+    backgroundColor: theme.primary + "20",
+    borderColor: theme.primary,
+    color: theme.primary,
+  },
+  qualityModifier: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    fontWeight: "400",
+  },
+  cookingControls: {
+    marginBottom: "20px",
+  },
+  cookButton: {
+    width: "100%",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    gap: "8px",
+    padding: "16px",
+    backgroundColor: theme.primary,
+    color: "white",
+    border: "none",
+    borderRadius: "8px",
+    fontSize: "16px",
+    fontWeight: "600",
+    cursor: "pointer",
+    transition: "all 0.2s ease",
+  },
+  cookButtonDisabled: {
+    backgroundColor: theme.textSecondary,
+    cursor: "not-allowed",
+  },
+  spinner: {
+    width: "16px",
+    height: "16px",
+    border: "2px solid transparent",
+    borderTop: "2px solid white",
+    borderRadius: "50%",
+    animation: "spin 1s linear infinite",
+  },
+  qualityEffects: {
+    marginBottom: "20px",
+  },
+  qualityList: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+  },
+  qualityEffect: {
+    padding: "12px",
+    backgroundColor: theme.background,
+    borderRadius: "8px",
+    border: `1px solid ${theme.border}`,
+  },
+  qualityEffectHeader: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    fontSize: "14px",
+    fontWeight: "600",
+    marginBottom: "8px",
+  },
+  qualityEffectText: {
+    fontSize: "13px",
+    color: theme.textSecondary,
+    lineHeight: "1.4",
+    margin: 0,
+  },
+  qualityFlawed: {
+    color: "#EF4444",
+  },
+  qualityRegular: {
+    color: "#6B7280",
+  },
+  qualityExceptional: {
+    color: "#3B82F6",
+  },
+  qualitySuperior: {
+    color: "#8B5CF6",
+  },
+  lastResultCard: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  resultHeader: {
+    marginBottom: "16px",
+  },
+  resultTitle: {
+    display: "flex",
+    alignItems: "center",
+    gap: "8px",
+    fontSize: "18px",
+    fontWeight: "600",
+    color: theme.text,
+    margin: 0,
+  },
+  resultContent: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "12px",
+  },
+  resultRow: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  resultLabel: {
+    fontSize: "14px",
+    fontWeight: "500",
+    color: theme.textSecondary,
+  },
+  resultValue: {
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  diceRoll: {
+    fontSize: "14px",
+    fontFamily: "monospace",
+    color: theme.text,
+  },
+  totalValue: {
+    fontWeight: "700",
+    color: theme.primary,
+  },
+  qualityResult: {
+    display: "flex",
+    alignItems: "center",
+    gap: "4px",
+    fontSize: "14px",
+    fontWeight: "600",
+    textTransform: "capitalize",
+  },
+  resultTimestamp: {
+    fontSize: "12px",
+    color: theme.textSecondary,
+    textAlign: "center",
+    fontStyle: "italic",
+    marginTop: "8px",
+  },
+  referenceGrid: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax(300px, 1fr))",
+    gap: "20px",
+    marginTop: "32px",
+  },
+  card: {
+    backgroundColor: theme.surface,
+    borderRadius: "12px",
+    padding: "20px",
+    border: `2px solid ${theme.border}`,
+    boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+  },
+  cardTitle: {
+    fontSize: "18px",
+    fontWeight: "600",
+    color: theme.text,
+    marginBottom: "16px",
+    margin: 0,
+  },
+  referenceTable: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "2px",
+  },
+  referenceTableHeader: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr",
+    gap: "12px",
+    padding: "8px 12px",
+    backgroundColor: theme.primary + "20",
+    borderRadius: "6px",
+    fontSize: "14px",
+    fontWeight: "600",
+    color: theme.text,
+  },
+  referenceTableRow: {
+    display: "grid",
+    gridTemplateColumns: "1fr 1fr",
+    gap: "12px",
+    padding: "8px 12px",
+    borderBottom: `1px solid ${theme.border}`,
+    fontSize: "14px",
+    color: theme.text,
+  },
+  referenceTableRowHeader: {
+    fontWeight: "600",
+    textTransform: "capitalize",
+  },
+  ingredientEffectsList: {
+    display: "flex",
+    flexDirection: "column",
+    gap: "8px",
+  },
+  ingredientEffectRow: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "8px 12px",
+    backgroundColor: theme.background,
+    borderRadius: "6px",
+    fontSize: "14px",
+  },
+  ingredientEffectLabel: {
+    fontWeight: "600",
+    textTransform: "capitalize",
+    color: theme.text,
+  },
+  ingredientEffectValue: {
+    fontWeight: "500",
+    color: theme.text,
+  },
+  ingredientEffectHarder: {
+    color: "#EF4444",
+  },
+  ingredientEffectEasier: {
+    color: "#10B981",
+  },
+});
+
+// Helper function for backward compatibility
+export const getRecipesStyles = (theme) => createRecipesStyles(theme);
